Style : 모달 코드 분리
- 모달 코드의 가독성 및 유지보수성을 위해서 코드 분리
