+++ src/components/Modal/style.ts
