+++ src/containers/DetailModal/index.tsx
@@ -1,23 +1,13 @@
 import React, { useMemo } from "react";
-import { ChampDetail } from "./style";
 
 import { useChampionDetail } from "@/apis";
 import { IChampionDetail } from "@/apis/lol/lol.model";
-import { useModal, useFilter, useCustomTheme } from "@/hooks";
 
-import { Cancel } from "@mui/icons-material";
-import {
-  Box,
-  Chip,
-  DialogContent,
-  IconButton,
-  Skeleton,
-  Typography,
-  useMediaQuery,
-  useTheme,
-} from "@mui/material";
+import { Modal } from "@/components";
+import { useModal, useFilter } from "@/hooks";
 
-type championData = {
+export type championData = {
+  champId: string;
   bgImgMb: string;
   bgImg: string;
   champName: string;
@@ -30,7 +20,6 @@ type DetailModalProps = {};
 
 export const DetailModal: React.FC<DetailModalProps> = ({}) => {
   const { closeModal, open, championId } = useModal();
-  const { isThemeLight } = useCustomTheme();
   const { language } = useFilter();
 
   const { data, isLoading } = useChampionDetail({
@@ -38,12 +27,10 @@ export const DetailModal: React.FC<DetailModalProps> = ({}) => {
     language: language,
   });
 
-  const screen = useTheme();
-  const isFullScreenXS = useMediaQuery(screen.breakpoints.down("sm")); // full screen의 breakpoints를 걸어주기 위한 코드
-
   const champData: championData = useMemo(() => {
     const champion = (data?.data?.[championId] as IChampionDetail) || {};
     return {
+      champId: championId,
       bgImgMb: `https://ddragon.leagueoflegends.com/cdn/img/champion/loading/${championId}_0.jpg`,
       bgImg: `https://ddragon.leagueoflegends.com/cdn/img/champion/splash/${championId}_0.jpg`,
       champName: champion.name || "",
@@ -53,102 +40,12 @@ export const DetailModal: React.FC<DetailModalProps> = ({}) => {
     };
   }, [championId, data]);
 
-  const handleClose = () => {
-    closeModal();
-  };
-
   return (
-    <ChampDetail
-      open={open}
-      onClose={handleClose}
-      fullWidth={true}
-      maxWidth={"lg"}
-      fullScreen={isFullScreenXS}
-    >
-      <DialogContent
-        className="dialog-content"
-        sx={{
-          backgroundImage: `url(${
-            isFullScreenXS ? champData.bgImgMb : champData.bgImg
-          })`,
-          height: { xs: "100vh", sm: "90vh", md: "80vh" },
-        }}
-      >
-        <Box
-          className="dialog-content-box"
-          sx={{
-            padding: {
-              xs: "20px 16px",
-              sm: "30px",
-              md: "50px 52px",
-              lg: "70px 72px",
-            },
-            background: isThemeLight
-              ? "linear-gradient(45deg, #ffffff, #ffffff42, #fff0)"
-              : "linear-gradient(45deg, #000000, #00000042, #fff0)",
-          }}
-        >
-          <IconButton
-            className="dialog-content-box-cancle"
-            size="large"
-            onClick={handleClose}
-            children={<Cancel />}
-          />
-
-          {isLoading ? (
-            <>
-              <Skeleton variant="rounded" width={"60%"} height={20} />
-              <Skeleton
-                variant="rounded"
-                width={"50%"}
-                height={50}
-                sx={{ mt: 2, mb: 2 }}
-              />
-              <Skeleton variant="text" width={"30%"} height={40} />
-              <Skeleton
-                variant="rounded"
-                sx={{ width: { xs: "100%", sm: "65%" }, mt: 2 }}
-                height={200}
-              />
-            </>
-          ) : (
-            <>
-              <Typography
-                className="dialog-content-box-subtitle"
-                variant="h5"
-                fontWeight={"regular"}
-                textAlign={"left"}
-              >
-                {champData?.champTitle}
-              </Typography>
-              <Typography variant="h2" fontWeight={"bold"} textAlign={"left"}>
-                {champData.champName}
-              </Typography>
-
-              {champData.champTags.length > 0 && (
-                <div className="dialog-content-box-tags">
-                  {champData.champTags?.map((tag: string) => (
-                    <Chip
-                      label={`#${tag}`}
-                      size="small"
-                      itemType="detail"
-                      key={`${championId}_${tag}`}
-                    />
-                  ))}
-                </div>
-              )}
-
-              <Typography
-                variant="body1"
-                textAlign={"left"}
-                sx={{ width: { xs: "100%", sm: "65%" }, mt: 2 }}
-              >
-                {champData.champLore}
-              </Typography>
-            </>
-          )}
-        </Box>
-      </DialogContent>
-    </ChampDetail>
+    <Modal
+      isOpen={open}
+      handleClose={closeModal}
+      isLoading={isLoading}
+      champData={champData}
+    />
   );
 };
